backend/biboon/pyproject.toml
[tool.poetry]
name = "biboon"
version = "0.1.0"
description = ""
authors = ["nnnLik <rslmakhmudov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic-settings = "2.2.1"
aio-pika = "9.4.1"
ujson = "5.9.0"
motor = "3.4.0"
beanie = "1.25.0"
sanic = "23.12.1"
sanic-cors = "2.2.0"
redis = {extras = ["hiredis"], version = "5.0.3"}
uvicorn = "0.29.0"


[tool.poetry.group.dev.dependencies]
black = "^24.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
